+++ packages/holz-json-backend/src/index.ts
@@ -1,0 +1,1 @@
+export { default, default as JsonBackend } from './json-backend';