+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- `[@holz/json-backend]` Initial release.
+
 ### Changed
 
 - `[@holz/stream-backend]` Use OS-sensitive line endings.